backendGit/requirements.txt
flask
flask-restful
flask-jwt-extended
flask-sqlalchemy
marshmallow_sqlalchemy